requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-telegram-bot>=20.0
APScheduler>=3.10.4
opencv-python>=4.8.0
pytesseract>=0.3.10
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.5
pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
chromadb>=0.4.0
google-generativeai>=0.7.2
soundfile>=0.12.1
SpeechRecognition>=3.10.0
langdetect>=1.0.9
deep-translator>=1.11.4